Kreirana klasa Podaci_grafovi
