+++ jknjk.suml
@@ -1,0 +1,2 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Classdiagram><settings><option name="paintImplements" value="true" /><option name="paintDepends" value="false" /><option name="layoutOnChanges" value="false" /><option name="paintUses" value="false" /><option name="paintInner" value="true" /><option name="paintExtends" value="true" /></settings><classes><class name="hr.foi.aplikacija.Trenutno_stanje" x="0" y="0"><option name="pinned" value="false" /><option name="fieldsExpanded" value="false" /><option name="constructorsExpanded" value="false" /><option name="methodsExpanded" value="false" /></class></classes><stickycomponents /><textcomponents /><connectors /></Classdiagram>
